move suggestion code to its own submodule
